reply function debugged

+++ imdbProject/src/main/java/com/example/imdbproject/repository/CommentRepository.java
@@ -5,9 +5,11 @@ import com.example.imdbproject.model.TitleBasic;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
 import java.util.Set;
 
 
-public interface CommentRepository extends JpaRepository<Comment,Integer> {
+public interface CommentRepository extends JpaRepository<Comment,Long> {
     Set<Comment> findByTitleBasic (TitleBasic titleBasic);
+
 }
